+ struct refactor + [subtask1] avoid List<> using
