+++ src/main/java/ru/tn/courses/dbatishchev/task1/v2/subtask1/Subtask1.java
@@ -1,12 +1,14 @@
-package ru.tn.courses.dbatishchev.task1;
+package ru.tn.courses.dbatishchev.task1.v2.subtask1;
 
-import java.util.ArrayList;
+import ru.tn.courses.dbatishchev.task1.v2.Aux;
+
+import java.util.Arrays;
 import java.util.List;
 
 /**
  * Created by ii343hbka on 01/03/2017.
  */
-public class Exercise1 {
+public class Subtask1 {
 
     private static final long seqSize = 100;
     private static final int seqLowBound = 0;
@@ -14,13 +16,13 @@ public class Exercise1 {
 
     public static void main(String[] args) {
         int[] intSeq = Aux.randomIntSequence(seqSize, seqLowBound, seqUpperBound);
-        Integer[] zeroElemIndices;
-        List<Integer> zeroElemIndicesList = new ArrayList<>();
+        int[] zeroElemIndices = new int[intSeq.length];
 
+        int zIdx = 0;
         for(int i = 0; i < intSeq.length; i++)
             if(intSeq[i] == 0)
-                zeroElemIndicesList.add(i);
-        zeroElemIndices = zeroElemIndicesList.toArray(new Integer[0]);
+                zeroElemIndices[zIdx++] = i;
+        zeroElemIndices = Arrays.copyOfRange(zeroElemIndices, 0, zIdx);
 
         System.out.println("Zero elements count: " + zeroElemIndices.length);
         System.out.print("Indices: ");
